new edit

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: mmondad <mmondad@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/31 14:33:13 by mmondad           #+#    #+#             */
-/*   Updated: 2023/11/03 17:49:11 by mmondad          ###   ########.fr       */
+/*   Updated: 2023/11/19 19:09:54 by mmondad          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
